+++ servant-server/src/Servant/Server/Internal/ResponseRender.hs
@@ -26,7 +26,7 @@ import Servant.API.ContentTypes
   )
 import Servant.API.MultiVerb
 import Servant.API.Status
-import Servant.API.Stream (SourceIO)
+import Servant.API.Stream (FramingRender (..), SourceIO)
 import Servant.API.UVerb.Union
 import Servant.Types.Internal.Response
 import qualified Servant.Types.SourceT as S
@@ -150,6 +150,24 @@ instance
       InternalResponse
         { statusCode = statusVal (Proxy @s)
         , responseBody = x
+        , headers = mempty
+        }
+
+instance
+  ( Accept ct
+  , FramingRender framing
+  , KnownStatus s
+  , MimeRender ct chunk
+  )
+  => ResponseRender cs (RespondStream s desc framing ct chunk)
+  where
+  type ResponseStatus (RespondStream s desc framing ct chunk) = s
+  type ResponseBody (RespondStream s desc framing ct chunk) = SourceIO ByteString
+  responseRender _ sourceChunks =
+    pure . addContentType @ct $
+      InternalResponse
+        { statusCode = statusVal (Proxy @s)
+        , responseBody = BSL.toStrict <$> framingRender (Proxy @framing) (mimeRender (Proxy @ct)) sourceChunks
         , headers = mempty
         }